+++ src/main/java/com/example/foodthought/dto/comment/CommentRequest.java
@@ -1,5 +1,6 @@
 package com.example.foodthought.dto.comment;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
