+++ src/main/java/com/example/foodthought/repository/CommentRepository.java
@@ -1,13 +1,15 @@
 package com.example.foodthought.repository;
 
 import com.example.foodthought.entity.Comment;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 //    List<Comment> findByBoardId(Long boardId);
-    List<Comment> findByBoardIdAndParentCommentIsNull(Long boardId);
+    Page<Comment> findByBoardIdAndParentCommentIsNull(Long boardId, Pageable pageable);
 
     List<Comment> findCommentsByBoard_Id(Long boardId);
 }
