+++ src/main/java/com/example/foodthought/service/AuthService.java
@@ -2,59 +2,24 @@ package com.example.foodthought.service;
 
 import com.example.foodthought.common.dto.ResponseDto;
 import com.example.foodthought.dto.auth.LoginUserDto;
-import com.example.foodthought.security.jwt.JwtUtil;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+public interface AuthService {
+    /**
+     * 로그인
+     * @param loginUserDto 로그인에 필요한 정보
+     * @param response 사용자에게 Token 전달
+     * @return 성공시 true, 실패는 exception 처리
+     */
+    public ResponseDto<Boolean> login(LoginUserDto loginUserDto, HttpServletResponse response);
 
-@RequiredArgsConstructor
-@Service
-public class AuthService {
 
-    private final AuthenticationManager authenticationManager;
-    private final JwtUtil jwtUtil;
-    public ResponseDto login(LoginUserDto loginUserDto,
-                             HttpServletResponse response){
-        Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(loginUserDto.getUserId(),loginUserDto.getPassword()));
-
-        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-        Collection<? extends GrantedAuthority> authorities = userDetails.getAuthorities();
-
-        List<String> userRols = new ArrayList<>();
-        for (GrantedAuthority authority : authorities) {
-            String userRole = authority.getAuthority();
-            userRols.add(userRole);
-        }
-
-        String token = jwtUtil.createToken(userDetails.getUsername(),userRols).split(" ")[1];
-        Cookie cookie = new Cookie(JwtUtil.AUTHORIZATION_HEADER,token);
-        cookie.setMaxAge((int)jwtUtil.TOKEN_TIME);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-
-        return ResponseDto.success(HttpStatus.OK.value());
-    }
-    public ResponseDto logout(HttpServletRequest request,
-                              HttpServletResponse response){
-
-        Cookie cookie = new Cookie(JwtUtil.AUTHORIZATION_HEADER,null);
-        cookie.setMaxAge(0);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-        return ResponseDto.success(HttpStatus.OK.value());
-    }
+    /**
+     * 로그인
+     * @param request 로그인한 이용자의 쿠키 보유
+     * @param response 초기화된 쿠키를 전달
+     * @return 성공시 true, 실패는 exception 처리
+     */
+    public ResponseDto<Boolean> logout(HttpServletRequest request, HttpServletResponse response);
 }
