+++ src/test/java/com/example/foodthought/entity/CommentTest.java
@@ -1,0 +1,48 @@
+package com.example.foodthought.entity;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class CommentTest {
+
+    @Test
+    void block() {
+    }
+
+    @Test
+    void updateComment() {
+    }
+
+    @Test
+    void getCommentId() {
+    }
+
+    @Test
+    void getContents() {
+    }
+
+    @Test
+    void getStatus() {
+    }
+
+    @Test
+    void getUser() {
+    }
+
+    @Test
+    void getBoard() {
+    }
+
+    @Test
+    void getParentComment() {
+    }
+
+    @Test
+    void getReplies() {
+    }
+
+    @Test
+    void builder() {
+    }
+}
